Javet v0.8.10 (#40)
* Renamed ``IJavetConsumer`` to ``IJavetUniConsumer``
* Added ``IJavetUniIndexedConsumer`` and ``IJavetBiIndexedConsumer``
* Fixed a bug in ``V8FunctionCallback`` on varargs
* Deprecated ``bindFunctions()`` and ``bindProperties()``
* Added ``@CheckReturnValue`` to warn ignored return value

+++ src/main/java/com/caoccao/javet/values/primitive/V8ValueDouble.java
@@ -47,6 +47,6 @@ public class V8ValueDouble extends V8ValuePrimitive<Double> {
     }
 
     public double toPrimitive() {
-        return value.doubleValue();
+        return value;
     }
 }
